Changes in HTML and script
HTML:
divide display into two parts - number_display and sign-display
(to show sign of operator) with <span>. The second argument (or result
of previous operation) and the operator should be visible
simultaneously.

Script:
According to changes in HTML, change number output to
numberDisplay instead display element.

Fix roundNumber function. Also now the maximal length of positive and
negative numbers is the same (because it can be changed by user).

Fix onclickDigitEvent function work with number length and change .pop
to .shift when deleting zero in the begining of the number (it works,
but it is ridiculous :-))